requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
dlib==19.24.2
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
grad-cam==1.5.0
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
sympy==1.12
threadpoolctl==3.3.0
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.2
ttach==0.0.3
typing_extensions==4.9.0
urllib3==2.2.1
opencv_python==4.7.0.72
pandas==1.5.2
Pillow==10.2.0
